requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
bcrypt==5.0.0
click==8.3.1
colorama==0.4.6
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.7.0
fastapi==0.128.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
mongoengine==0.27.0
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.6.1
pytz==2025.2
SQLAlchemy==2.0.45
sqlparse==0.2.4
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
whitenoise==6.11.0
openpyxl==3.1.2
